CIのスクリプトをまとめる
