+++ .github/workflows/format.yml
@@ -29,8 +29,7 @@ jobs:
           node-version-file: .node-version
           cache: npm
           cache-dependency-path: package-lock.json
-      - run: bash "${GITHUB_WORKSPACE}/scripts/format/format/npm_install.sh"
-      - run: npm run fix
+      - run: bash "${GITHUB_WORKSPACE}/scripts/format/format/fix.sh"
       - uses: dev-hato/actions-diff-pr-management@v1.0.8
         with:
           github-token: ${{secrets.GITHUB_TOKEN}}
