+++ .github/workflows/renovate-config-validator.yml
@@ -17,8 +17,7 @@ jobs:
         with:
           node-version-file: .node-version
           cache: npm
-      - run: bash "${GITHUB_WORKSPACE}/scripts/npm_ci.sh"
-      - run: npx renovate-config-validator
+      - run: bash "${GITHUB_WORKSPACE}/scripts/renovate_config_validator/renovate_config_validator/validate.sh"
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.event_name }}-${{ github.ref }}
