+++ scripts/cdk_deploy/cdk_deploy/deploy.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+bash "${GITHUB_WORKSPACE}/scripts/npm_ci.sh"
+npx cdk deploy --require-approval never
