--- scripts/docker_push/dockle/install_dockle.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-dockle_version="$(cat .dockle-version)"
-curl -L -o dockle.deb "https://github.com/goodwithtech/dockle/releases/download/v${dockle_version}/dockle_${dockle_version}_Linux-64bit.deb"
-sudo dpkg -i dockle.deb
