+++ scripts/renovate_config_validator/renovate_config_validator/validate.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+bash "${GITHUB_WORKSPACE}/scripts/npm_ci.sh"
+npx renovate-config-validator
