+++ scripts/update_major_node/update_major_node/update.js
@@ -1,5 +1,6 @@
 const fs = require('fs')
 const yaml = require('js-yaml')
+const renovate = require(`${process.env.GITHUB_WORKSPACE}/renovate.json`)
 
 module.exports = () => {
   const configFilename = `${process.env.GITHUB_WORKSPACE}/.github/dependabot.yml`
@@ -17,6 +18,19 @@ module.exports = () => {
 
   try {
     fs.writeFileSync(configFilename, '---\n' + yaml.dump(config), 'utf8')
+  } catch (err) {
+    console.error(err.message)
+    process.exit(1)
+  }
+
+  for (const rule of renovate.packageRules) {
+    if (rule.matchPackageNames.includes('@types/node')) {
+      rule.allowedVersions = `<=${process.env.LAMBDA_VERSION}`
+    }
+  }
+
+  try {
+    fs.writeFileSync(`${process.env.GITHUB_WORKSPACE}/renovate.json`, JSON.stringify(renovate, null, '  ') + '\n', 'utf8')
   } catch (err) {
     console.error(err.message)
     process.exit(1)
