+++ scripts/update_packages/update_packages/update.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+yq -i -o json '.dependencies.express|="~${{ steps.get_types_express_version.outputs.result }}"' package.json
+npm install
